[refactor] Response 분리
