+++ core/network/src/main/java/com/moya/funch/network/dto/response/match/ChemistryResponse.kt
@@ -1,0 +1,12 @@
+package com.moya.funch.network.dto.response.match
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ChemistryResponse(
+    @SerialName("title")
+    val title: String = "",
+    @SerialName("description")
+    val description: String = "",
+)
