+++ core/network/src/main/java/com/moya/funch/network/dto/response/match/RecommendResponse.kt
@@ -1,0 +1,10 @@
+package com.moya.funch.network.dto.response.match
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RecommendResponse(
+    @SerialName("title")
+    val title: String = "",
+)
