+++ core/network/src/main/java/com/moya/funch/network/dto/response/match/SubwayResponse.kt
@@ -1,0 +1,12 @@
+package com.moya.funch.network.dto.response.match
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SubwayResponse(
+    @SerialName("lines")
+    val lines: List<String> = listOf(),
+    @SerialName("name")
+    val name: String = "",
+)
