+++ core/network/src/main/java/com/moya/funch/network/dto/response/profile/ProfileResponse.kt
@@ -1,0 +1,20 @@
+package com.moya.funch.network.dto.response.profile
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ProfileResponse(
+    @SerialName("name")
+    val name: String = "",
+    @SerialName("jobGroup")
+    val jobGroup: String = "",
+    @SerialName("clubs")
+    val clubs: List<String> = listOf(),
+    @SerialName("mbti")
+    val mbti: String = "",
+    @SerialName("constellation")
+    val constellation: String = "",
+    @SerialName("subwayNames")
+    val subwayNames: List<String> = listOf(),
+)
