+++ core/network/src/main/java/com/moya/funch/network/service/MatchingService.kt
@@ -2,7 +2,7 @@ package com.moya.funch.network.service
 
 import com.moya.funch.network.dto.request.MatchingRequest
 import com.moya.funch.network.dto.response.BaseResponse
-import com.moya.funch.network.dto.response.MatchingResponse
+import com.moya.funch.network.dto.response.match.MatchingResponse
 import retrofit2.http.Body
 import retrofit2.http.POST
